+++ frontend/src/components/Button/index.js
@@ -1,3 +1,4 @@
 export { default as BasicButton } from "./basicButton.js";
 export { default as BackIconButton } from "./backIconButton.js";
 export { default as ModeButton } from "./modeButton.js";
+export { default as IconButton } from "./iconButton.js";
