+++ frontend/src/components/Options/Account/exitOption.js
@@ -1,4 +1,5 @@
-import { Component } from "../../../core/index.js";
+import { Component, createComponent } from "../../../core/index.js";
+import { IconButton } from "../../Button/index.js";
 
 class ExitOption extends Component {
   render() {
@@ -10,18 +11,12 @@ class ExitOption extends Component {
       "text-center text-primary-text text-[40px] font-semibold font-['Inter'] leading-10";
     optionText.textContent = this.props.text;
 
-    const iconContainer = document.createElement("div");
-    iconContainer.className = "w-[72px] h-[72px] relative";
+    const iconContainer = createComponent(IconButton, {
+      iconSrc: this.props.iconSrc,
+      bgColorClass: this.props.bgColorClass,
+      optionName: this.props.optionName,
+    });
 
-    const iconBg = document.createElement("div");
-    iconBg.className = `absolute inset-0 m-auto flex items-center justify-center ${this.props.bgColorClass} rounded-full`;
-
-    const iconImage = document.createElement("img");
-    iconImage.src = this.props.iconSrc;
-    iconImage.className = "w-[48px] h-[48px]";
-
-    iconBg.appendChild(iconImage);
-    iconContainer.appendChild(iconBg);
     option.appendChild(optionText);
     option.appendChild(iconContainer);
